move packetparser to dev_server
